reverted User to interface, created UserService
