+++ src/config/clientConfig.ts
@@ -1,0 +1,8 @@
+export default {
+  apiKey: "AIzaSyDCDeYDMBWkNDv6vYHvliK75FZIYeui8go",
+  authDomain: "catholic-gators-dev.firebaseapp.com",
+  databaseURL: "https://catholic-gators-dev.firebaseio.com",
+  projectId: "catholic-gators-dev",
+  storageBucket: "catholic-gators-dev.appspot.com",
+  messagingSenderId: "1019003476611"
+};
