+++ src/database/firestore.ts
@@ -1,48 +1,25 @@
 import { app, auth } from 'firebase/app';
 import 'firebase/firestore';
 import 'firebase/auth';
-import { Observable, from, merge } from 'rxjs';
-import { mergeMap, map, partition } from 'rxjs/operators';
+import { Observable, from } from 'rxjs';
+import { map } from 'rxjs/operators';
 
-import User from '../models/user';
 import Document from '../models/document';
 import DocumentNotFoundError from '../models/documentNotFoundError';
 
-const USER_COLLECTION = 'users';
-const ADMIN_EMAIL = 'sandy@catholicgators.org'; // TODO: we might want to migrate this to some form of config
-
 export default class Firestore {
     private app: app.App;
-    private auth: auth.Auth;
     private db: firebase.firestore.Firestore;
-    private firebaseUser$: Observable<firebase.User>;
-    private user$: Observable<User>;
 
     constructor(private firebase, private clientConfig) {
         this.app = !this.firebase.apps.length ?
                         this.firebase.initializeApp(this.clientConfig) :
                         this.firebase.app();
-        this.auth = this.app.auth();
         this.db = this.app.firestore();
-        this.firebaseUser$ = Observable.create(observer =>
-            this.auth.onAuthStateChanged(
-                user => observer.next(user),
-                err => observer.error(err)
-            )
-        );
-        this.user$ = this.storeUserInfoOnLogin();
     };
 
-    listenForUser(): Observable<User> {
-        return this.user$;
-    }
-
-    googleSignIn(): Observable<void> {
-        return from(this.auth.signInWithRedirect(new auth.GoogleAuthProvider()));
-    }
-
-    signOut(): Observable<void> {
-        return from(this.auth.signOut());
+    getAuth(): auth.Auth {
+        return this.app.auth();
     }
 
     addDoc(collection: string, entity: object): Observable<firebase.firestore.DocumentReference> {
@@ -86,34 +63,6 @@ export default class Firestore {
             );
     }
 
-    getUsers() {
-        return this.getCollection(USER_COLLECTION);
-    }
-
-    approveUser(id: string) {
-        this.db.collection(USER_COLLECTION).doc(id).update({
-            isApproved: true
-        });
-    }
-
-    disapproveUser(id: string) {
-        this.db.collection(USER_COLLECTION).doc(id).update({
-            isApproved: false
-        });
-    }
-
-    makeAdmin(id: string) {
-        this.db.collection(USER_COLLECTION).doc(id).update({
-            isAdmin: true
-        });
-    }
-
-    removeAdmin(id: string) {
-        this.db.collection(USER_COLLECTION).doc(id).update({
-            isAdmin: false
-        });
-    }
-
     doesExist(collection: string, docId: string): Observable<boolean> {
         return from(this.db.collection(collection).doc(docId).get())
             .pipe(
@@ -133,54 +82,4 @@ export default class Firestore {
         return from(this.app.delete());
     }
 
-    private storeUserInfoOnLogin(): Observable<User> {
-        const attachExistence = firebaseUser =>
-                    this.doesExist(USER_COLLECTION, firebaseUser.uid)
-                        .pipe(
-                            map(exists => ({exists, firebaseUser}))
-                        );
-
-        const storeUser = ({exists, firebaseUser}) => {
-                    const docId = firebaseUser.uid;
-                    const user: User = this.wrapFirebaseUser(firebaseUser);
-                    let observable;
-
-                    if (exists) {
-                        // If the user already exists, just update their profile pic
-                        observable = this.updateDoc(USER_COLLECTION, docId, {
-                            photoURL: user.photoURL
-                        })
-                    } else {
-                        observable = this.upsertDocById(USER_COLLECTION, docId, user);
-                    }
-
-                    return observable.pipe(
-                        mergeMap(_ => this.getDoc(USER_COLLECTION, firebaseUser.uid)),
-                        map(userObject => (<Document>userObject).data)
-                    ) as Observable<User>;
-                }
-
-        let [login$, logout$] = partition(x => x !== null)(this.firebaseUser$);
-        login$ = login$
-            .pipe(
-                mergeMap(attachExistence),
-                mergeMap(storeUser)
-        );
-
-        return merge(
-            login$,
-            logout$,
-        ) as Observable<User>;
-    }
-
-    private wrapFirebaseUser(user: firebase.User): User {
-        return {
-            name: user.displayName,
-            email: user.email,
-            photoURL: user.photoURL,
-            isApproved: user.email === ADMIN_EMAIL,
-            isAdmin: user.email === ADMIN_EMAIL,
-        }
-
-    }
 }
